Add files via upload
This is the main appication rar.. Download it and extract to any location.. Then open cmd prompt goto the path of the folder where you have kept it. Then type "npm install", let the download complete. Then type "npm start".... It will redirect to default browser of with the login page.